# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Delivery API",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o token JWT no formato: Bearer {token}"
    }
  },
...
  "paths": {
    "/stock": {
      "get": {
        "tags": [
          "Estoque"
        ],
        "summary": "Listar estoque da empresa logada",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de produtos da empresa com estoque"
          },
          "403": {
            "description": "Acesso negado"
          }
        }
      }
    },
    "/stock/{productId}": {
      "patch": {
        "tags": [
          "Estoque"
        ],
        "summary": "Atualizar quantidade de estoque de um produto",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do produto"
          },
          {
            "in": "body",
            "name": "dados",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "estoque": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "required": [
                "estoque"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estoque atualizado com sucesso"
          },
          "400": {
            "description": "Estoque inválido"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Estoque"
        ],
        "summary": "Remover produto do estoque (da empresa)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto removido com sucesso"
          },
          "404": {
            "description": "Produto não encontrado para esta empresa"
          }
        }
      }
    },
    "/staff": {
      "get": {
        "tags": [
          "Funcionários"
        ],
        "summary": "Listar todos os funcionários da empresa logada",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de funcionários",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "nome": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "telefone": {
                    "type": "string"
                  },
                  "cargo": {
                    "type": "string"
                  },
                  "placaVeiculo": {
                    "type": "string"
                  },
                  "empresaId": {
                    "type": "integer"
                  },
                  "ativo": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado (usuário não é empresa)"
          }
        }
      },
      "post": {
        "tags": [
          "Funcionários"
        ],
        "summary": "Cadastrar novo funcionário",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "funcionario",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "telefone": {
                  "type": "string"
                },
                "cargo": {
                  "type": "string",
                  "description": "Exemplo: motoboy, cozinheiro, gerente"
                },
                "placaVeiculo": {
                  "type": "string",
                  "description": "Campo opcional, necessário apenas para motoboys"
                }
              },
              "required": [
                "nome",
                "email",
                "cargo"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Funcionário cadastrado com sucesso"
          },
          "400": {
            "description": "Campos obrigatórios ausentes"
          }
        }
      }
    },
    "/staff/{id}": {
      "patch": {
        "tags": [
          "Funcionários"
        ],
        "summary": "Atualizar dados de um funcionário",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "dadosAtualizados",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "telefone": {
                  "type": "string"
                },
                "cargo": {
                  "type": "string"
                },
                "placaVeiculo": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Funcionário atualizado com sucesso"
          },
          "404": {
            "description": "Funcionário não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Funcionários"
        ],
        "summary": "Desativar funcionário",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Funcionário desativado com sucesso"
          },
          "404": {
            "description": "Funcionário não encontrado"
          }
        }
      }
    },
...
    "/reviews": {
      "post": {
        "tags": [
          "Avaliações"
        ],
        "summary": "Enviar uma avaliação após entrega",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "review",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "orderId": {
                  "type": "string"
                },
                "companyId": {
                  "type": "integer"
                },
                "rating": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 5
                },
                "comment": {
                  "type": "string"
                }
              },
              "required": [
                "orderId",
                "companyId",
                "rating"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Avaliação registrada com sucesso"
          },
          "400": {
            "description": "Erro de validação ou já avaliado"
          }
        }
      }
    },
    "/reviews/company/{companyId}": {
      "get": {
        "tags": [
          "Avaliações"
        ],
        "summary": "Listar avaliações públicas de uma empresa",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Avaliações da empresa"
          }
        }
      }
    },
    "/reviews/my": {
      "get": {
        "tags": [
          "Avaliações"
        ],
        "summary": "Listar minhas avaliações",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Avaliações do usuário logado"
          }
        }
      }
    },
    "/reviews/my-company": {
      "get": {
        "tags": [
          "Avaliações"
        ],
        "summary": "Listar avaliações recebidas pela empresa logada",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Avaliações da empresa retornadas com sucesso"
          },
          "403": {
            "description": "Acesso negado (não é empresa)"
          }
        }
      }
    },
...
    "/products/{id}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Buscar produto por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        }
      },
      "patch": {
        "tags": [
          "Produtos"
        ],
        "summary": "Atualizar produto por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "dados",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto atualizado"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Produtos"
        ],
        "summary": "Remover produto por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto removido com sucesso"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        }
      }
    }
  }
}